+++ src/testFeatures/kotlin/dev/codebandits/container/gradle/DockerRemoveTest.kt
@@ -1,0 +1,97 @@
+package dev.codebandits.container.gradle
+
+import dev.codebandits.container.gradle.helpers.appendLine
+import org.gradle.testkit.runner.GradleRunner
+import org.gradle.testkit.runner.TaskOutcome
+import org.junit.jupiter.api.Test
+import strikt.api.expectThat
+import strikt.assertions.contains
+import strikt.assertions.isEqualTo
+import strikt.assertions.isFalse
+import strikt.assertions.isNotNull
+import java.util.*
+
+class DockerRemoveTest : GradleProjectTest() {
+
+  @Test
+  fun `dockerRemove removes the specified image`() {
+    pullImage("alpine:3.18.9")
+
+    buildGradleKtsFile.appendLine(
+      """
+      import dev.codebandits.container.gradle.tasks.ContainerRunTask
+      
+      plugins {
+        id("dev.codebandits.container")
+      }
+      
+      tasks {
+        register<ContainerRunTask>("removeAlpineImage") {
+          dockerRemove {
+            image = "alpine:3.18.9"
+          }
+        }
+      }
+      """.trimIndent()
+    )
+
+    val result = GradleRunner.create()
+      .withPluginClasspath()
+      .withProjectDir(projectDirectory.toFile())
+      .withArguments("removeAlpineImage")
+      .build()
+
+    expectThat(result).and {
+      get { task(":removeAlpineImage") }.isNotNull().get { outcome }.isEqualTo(TaskOutcome.SUCCESS)
+    }
+
+    expectThat(imageExists("alpine:3.18.9")).isFalse()
+  }
+
+  @Test
+  fun `dockerRemove fails when removing an image that does not exist`() {
+    buildGradleKtsFile.appendLine(
+      """
+      import dev.codebandits.container.gradle.tasks.ContainerRunTask
+      
+      plugins {
+        id("dev.codebandits.container")
+      }
+      
+      tasks {
+        register<ContainerRunTask>("removeImageNotExist") {
+          dockerRemove {
+            image = "alpine:${UUID.randomUUID()}"
+          }
+        }
+      }
+      """.trimIndent()
+    )
+
+    val result = GradleRunner.create()
+      .withPluginClasspath()
+      .withProjectDir(projectDirectory.toFile())
+      .withArguments("removeImageNotExist")
+      .buildAndFail()
+
+    expectThat(result).and {
+      get { task(":removeImageNotExist") }.isNotNull().get { outcome }.isEqualTo(TaskOutcome.FAILED)
+      get { output }.contains("No such image")
+    }
+  }
+
+  private fun pullImage(imageReference: String) {
+    ProcessBuilder("docker", "pull", imageReference)
+      .inheritIO()
+      .start()
+      .waitFor()
+  }
+
+  private fun imageExists(imageReference: String): Boolean {
+    val process = ProcessBuilder("docker", "image", "inspect", imageReference)
+      .redirectErrorStream(true)
+      .start()
+    val exitCode = process.waitFor()
+    return exitCode == 0
+  }
+}