+++ CHANGELOG.md
@@ -9,11 +9,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.3] - 2025-04-22
+## [Unreleased]
+
+### Added
+
+- dockerPull and dockerRemove
 
 ### Fixed
 
-- Remove the unreliable dependency on Docker CLI
+- Remove the unreliable dependency on Docker CLI. With this change, dockerRun no longer automatically pulls the image.
 
 ### Chore